Move route plugins to own folder and move initialization of route plugins to the bottle initialization module.
