+++ components/server/src/initialization/bottle.py
@@ -1,11 +1,13 @@
 """Initialize bottle."""
 
 import bottle
+from pymongo.database import Database
 
-from ..route_authentication_plugin import AuthenticationPlugin
+from ..route_plugins import AuthenticationPlugin, InjectionPlugin
 
 
-def init_bottle() -> None:
+def init_bottle(database: Database) -> None:
     """Initialize bottle."""
     bottle.BaseRequest.MEMFILE_MAX = 1024 * 1024  # Max size of POST body in bytes
+    bottle.install(InjectionPlugin(value=database, keyword="database"))
     bottle.install(AuthenticationPlugin())
