+++ components/server/src/route_plugins/__init__.py
@@ -1,0 +1,4 @@
+"""Route plugins."""
+
+from .route_authentication_plugin import AuthenticationPlugin
+from .route_injection_plugin import InjectionPlugin
