+++ components/server/tests/unittests/route_plugins/test_route_injection_plugin.py
@@ -4,8 +4,7 @@ import unittest
 
 import bottle
 
-from src.route_injection_plugin import InjectionPlugin
-from src.route_authentication_plugin import AuthenticationPlugin
+from src.route_plugins import AuthenticationPlugin, InjectionPlugin
 
 
 class RouteInjectionPluginTest(unittest.TestCase):
